+ @antonymarcano - consolidating the application elements into cohesive components
