+++ src/main/java/net/serenitybdd/demos/todos/pages/ApplicationHomePage.java
@@ -5,6 +5,4 @@ import net.serenitybdd.screenplay.targets.Target;
 import net.thucydides.core.annotations.DefaultUrl;
 
 @DefaultUrl("http://todomvc.com/examples/angularjs/#/")
-public class ApplicationHomePage extends PageObject {
-    public static final Target TITLE = Target.the("application title").locatedBy("header h1");
-}
+public class ApplicationHomePage extends PageObject {}
