--- src/main/java/net/serenitybdd/demos/todos/pages/todolist/ClearCompleted.java
@@ -1,7 +1,0 @@
-package net.serenitybdd.demos.todos.pages.todolist;
-
-import net.serenitybdd.screenplay.targets.Target;
-
-public class ClearCompleted {
-    public static Target BUTTON = Target.the("Clear completed button").locatedBy("#clear-completed");
-}
