--- src/main/java/net/serenitybdd/demos/todos/pages/todolist/ToDoList.java
@@ -1,7 +1,0 @@
-package net.serenitybdd.demos.todos.pages.todolist;
-
-import net.serenitybdd.screenplay.targets.Target;
-
-public class ToDoList {
-    public static Target TODO_ITEMS = Target.the("Todo items").locatedBy(".view label");
-}
