--- src/main/java/net/serenitybdd/demos/todos/pages/todolist/counter/TodoCounter.java
@@ -1,7 +1,0 @@
-package net.serenitybdd.demos.todos.pages.todolist.counter;
-
-import net.serenitybdd.screenplay.targets.Target;
-
-public class TodoCounter {
-    public static Target ITEM_COUNT = Target.the("todo count").locatedBy("#todo-count strong");
-}
