--- src/main/java/net/serenitybdd/demos/todos/pages/todolist/newitem/NewTodoForm.java
@@ -1,7 +1,0 @@
-package net.serenitybdd.demos.todos.pages.todolist.newitem;
-
-import net.serenitybdd.screenplay.targets.Target;
-
-public class NewTodoForm {
-    public static Target NEW_TODO_FIELD = Target.the("New Todo Field").locatedBy("#new-todo");
-}
