+++ src/main/java/net/serenitybdd/demos/todos/questions/DisplayedItems.java
@@ -1,6 +1,6 @@
 package net.serenitybdd.demos.todos.questions;
 
-import net.serenitybdd.demos.todos.pages.todolist.ToDoList;
+import net.serenitybdd.demos.todos.pages.ToDoList;
 import net.serenitybdd.screenplay.Actor;
 import net.serenitybdd.screenplay.Question;
 import net.serenitybdd.screenplay.annotations.Subject;
@@ -13,6 +13,6 @@ public class DisplayedItems implements Question<List<String>> {
 
     @Override
     public List<String> answeredBy(Actor actor) {
-        return Text.of(ToDoList.TODO_ITEMS).viewedBy(actor).asList();
+        return Text.of(ToDoList.ITEMS).viewedBy(actor).asList();
     }
 }
