+++ src/main/java/net/serenitybdd/demos/todos/questions/ItemsLeftCounter.java
@@ -1,6 +1,6 @@
 package net.serenitybdd.demos.todos.questions;
 
-import net.serenitybdd.demos.todos.pages.todolist.counter.TodoCounter;
+import net.serenitybdd.demos.todos.pages.ToDoList;
 import net.serenitybdd.screenplay.Actor;
 import net.serenitybdd.screenplay.Question;
 import net.serenitybdd.screenplay.questions.Text;
@@ -8,7 +8,7 @@ import net.serenitybdd.screenplay.questions.Text;
 public class ItemsLeftCounter implements Question<Integer> {
     @Override
     public Integer answeredBy(Actor actor) {
-        return Text.of(TodoCounter.ITEM_COUNT)
+        return Text.of(ToDoList.ITEMS_LEFT)
                    .viewedBy(actor)
                    .asInteger();
     }
