+++ src/main/java/net/serenitybdd/demos/todos/tasks/ClearCompletedItems.java
@@ -1,6 +1,6 @@
 package net.serenitybdd.demos.todos.tasks;
 
-import net.serenitybdd.demos.todos.pages.todolist.ClearCompleted;
+import net.serenitybdd.demos.todos.pages.ToDoList;
 import net.serenitybdd.screenplay.Actor;
 import net.serenitybdd.screenplay.Task;
 import net.serenitybdd.screenplay.actions.Click;
@@ -10,6 +10,6 @@ public class ClearCompletedItems implements Task {
 
     @Step("{0} clears all the completed items")
     public <T extends Actor> void performAs(T actor) {
-        actor.attemptsTo(Click.on(ClearCompleted.BUTTON));
+        actor.attemptsTo(Click.on(ToDoList.CLEAR_COMPLETED));
     }
 }
